requirements.txt
beautifulsoup4==4.12.2
CairoSVG==2.7.1
Jinja2==3.1.2
Pillow==9.5.0
Pillow==10.0.1
PyYAML==6.0
PyYAML==6.0.1
Requests==2.31.0
selenium==4.12.0
tqdm==4.65.0
webdriver_manager==4.0.0
